added asTrimmedString and  asSafeInt to utils.ts
